work_inventory/requirements.txt
appdirs==1.4.4
appnope==0.1.0
argon2-cffi==20.1.0
asgiref==3.3.1
attrs==20.1.0
backcall==0.2.0
bleach==3.1.5
bst==0.5.0
cachetools==4.2.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
city==0.1
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.5
django-cors-headers==3.7.0
django-import-export==2.5.0
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
google-api-core==1.25.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
hupper==1.10.2
idna==2.10
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
linkedlist==2.0.0
MarkupPy==1.14
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
nose==1.3.7
notebook==6.1.3
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.6
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
PasteDeploy==2.1.0
pexpect==4.8.0
pickleshare==0.7.5
plaster==1.0
plaster-pastedeploy==0.7
player==0.6.1
prometheus-client==0.8.0
prompt-toolkit==3.0.6
protobuf==3.14.0
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pygame==2.0.1
Pygments==2.6.1
pyparsing==2.4.7
pyramid==1.10.4
pyramid-jinja2==2.8
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.4.1
pyzmq==19.0.2
Random-Word==1.0.4
requests==2.24.0
requests-oauthlib==1.3.0
rise==5.6.1
rsa==4.7
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
translationstring==1.4
uritemplate==3.0.1
urllib3==1.25.10
venusian==3.0.0
virtualenv==20.4.0
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.6
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
zope.deprecation==4.4.0
zope.interface==5.1.0
